chore: move next base fee to header impl
