+++ crates/payload/builder/Cargo.toml
@@ -13,7 +13,6 @@ reth-primitives = { path = "../../primitives" }
 reth-rpc-types = { path = "../../rpc/rpc-types" }
 reth-rlp = { path = "../../rlp" }
 reth-interfaces = { path = "../../interfaces" }
-reth-consensus-common = { path = "../../consensus/common" }
 reth-revm-primitives = { path = "../../revm/revm-primitives" }
 
 ## ethereum
